+++ ipwhois/tests/test_whois.py
@@ -56,13 +56,13 @@ class TestWhois(TestCommon):
         # groups are messed up.
         tmp_dict = RIR_WHOIS['arin']['fields']
         tmp_dict['name'] = r'(NetName):[^\S\n]+(?P<val1>.+?)\n'
-        obj._parse_fields(
+        obj.parse_fields(
             response="\nNetName:        TEST\n",
             fields_dict=tmp_dict,
             dt_format=RIR_WHOIS['arin']['dt_format']
         )
 
-        obj._parse_fields(
+        obj.parse_fields(
             response="\nUpdated:        2012-02-24\n",
             fields_dict=RIR_WHOIS['arin']['fields'],
             dt_format=RIR_WHOIS['arin']['dt_format']
@@ -70,13 +70,13 @@ class TestWhois(TestCommon):
 
         log.debug('Testing field parse error. This should be followed by a '
                   'debug log.')
-        obj._parse_fields(
+        obj.parse_fields(
             response='\nUpdated:        2012-02-244\n',
             fields_dict=RIR_WHOIS['arin']['fields'],
             dt_format=RIR_WHOIS['arin']['dt_format']
         )
 
-    def test__get_nets_arin(self):
+    def test_get_nets_arin(self):
 
         net = Net('74.125.225.229')
         obj = Whois(net)
@@ -90,9 +90,9 @@ class TestWhois(TestCommon):
             '\nNetRange:       74.125.1.0 - 74.125.1.0'
             '\n'
         )
-        obj._get_nets_arin(multi_net_response)
+        obj.get_nets_arin(multi_net_response)
 
-    def test__get_nets_lacnic(self):
+    def test_get_nets_lacnic(self):
 
         net = Net('200.57.141.161')
         obj = Whois(net)
@@ -103,9 +103,9 @@ class TestWhois(TestCommon):
             '\ninetnum:     200.57.256/19\r\n'
             '\n'
         )
-        obj._get_nets_lacnic(multi_net_response)
+        obj.get_nets_lacnic(multi_net_response)
 
-    def test__get_nets_other(self):
+    def test_get_nets_other(self):
 
         net = Net('210.107.73.73')
         obj = Whois(net)
@@ -116,4 +116,4 @@ class TestWhois(TestCommon):
             '\ninetnum:        210.107.0.0 - 210.107.127.256\n'
             '\n'
         )
-        obj._get_nets_other(multi_net_response)
+        obj.get_nets_other(multi_net_response)