+++ ipwhois/whois.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2013-2017 Philip Hane
+# Copyright (c) 2013-2019 Philip Hane
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -324,16 +324,6 @@ class Whois:
 
         return ret
 
-    def _parse_fields(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('Whois._parse_fields() has been deprecated and will be '
-             'removed. You should now use Whois.parse_fields().')
-        return self.parse_fields(*args, **kwargs)
-
     def get_nets_arin(self, response):
         """
         The function for parsing network blocks from ARIN whois data.
@@ -415,16 +405,6 @@ class Whois:
 
         return nets
 
-    def _get_nets_arin(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('Whois._get_nets_arin() has been deprecated and will be '
-             'removed. You should now use Whois.get_nets_arin().')
-        return self.get_nets_arin(*args, **kwargs)
-
     def get_nets_lacnic(self, response):
         """
         The function for parsing network blocks from LACNIC whois data.
@@ -494,16 +474,6 @@ class Whois:
                 pass
 
         return nets
-
-    def _get_nets_lacnic(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('Whois._get_nets_lacnic() has been deprecated and will be '
-             'removed. You should now use Whois.get_nets_lacnic().')
-        return self.get_nets_lacnic(*args, **kwargs)
 
     def get_nets_other(self, response):
         """
@@ -576,16 +546,6 @@ class Whois:
                 pass
 
         return nets
-
-    def _get_nets_other(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('Whois._get_nets_other() has been deprecated and will be '
-             'removed. You should now use Whois.get_nets_other().')
-        return self.get_nets_other(*args, **kwargs)
 
     def lookup(self, inc_raw=False, retry_count=3, response=None,
                get_referral=False, extra_blacklist=None,
@@ -740,7 +700,7 @@ class Whois:
 
                     results['raw_referral'] = response_ref
 
-                temp_rnet = self._parse_fields(
+                temp_rnet = self.parse_fields(
                     response_ref,
                     RWHOIS['fields'],
                     field_list=field_list
@@ -758,15 +718,15 @@ class Whois:
 
         if asn_data['asn_registry'] == 'arin':
 
-            nets_response = self._get_nets_arin(response)
+            nets_response = self.get_nets_arin(response)
 
         elif asn_data['asn_registry'] == 'lacnic':
 
-            nets_response = self._get_nets_lacnic(response)
+            nets_response = self.get_nets_lacnic(response)
 
         else:
 
-            nets_response = self._get_nets_other(response)
+            nets_response = self.get_nets_other(response)
 
         nets.extend(nets_response)
 
@@ -788,7 +748,7 @@ class Whois:
 
                 dt_format = None
 
-            temp_net = self._parse_fields(
+            temp_net = self.parse_fields(
                 response,
                 RIR_WHOIS[asn_data['asn_registry']]['fields'],
                 section_end,